DCPS-227-(添加获取医院信息缓存)-ck

+++ hc-labmanagement-ms/src/main/java/com/hc/service/HospitalequimentServiceImpl.java
@@ -21,6 +21,7 @@ import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.util.ObjectUtils;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -50,11 +51,6 @@ public class HospitalequimentServiceImpl implements HospitalequimentService {
         String equipmenttypeid = hospitalEquimentTypeCommand.getEquipmenttypeid();
         if(StringUtils.isBlank(equipmenttypeid)){
             throw new IedsException(HospitalequimentEnumErrorCode.HOSPITAL_TYPE_ID_NOT_NULL.getCode());
-        }
-        //判断该医院该设备类型是不是被绑定过
-        HospitalequimentDTO hospitalequiment =hospitalequimentRepository.selectHospitalEquiment(hospitalcode,equipmenttypeid);
-        if (null!=hospitalequiment){
-            throw   new IedsException(HospitalequimentEnumErrorCode.THE_SAME_DEVICE_TYPE_EXISTS.getCode());
         }
         WorkTimeBlockCommand[] workTimeBlock = hospitalEquimentTypeCommand.getWorkTimeBlock();
         HospitalequimentDTO hospitalequimentDTO = BeanConverter.convert(hospitalEquimentTypeCommand, HospitalequimentDTO.class);
@@ -128,6 +124,12 @@ public class HospitalequimentServiceImpl implements HospitalequimentService {
         if (CollectionUtils.isNotEmpty(equipment)){
             throw  new IedsException(HospitalequimentEnumErrorCode.DEVICES_EXIST_UNDER_THIS_DEVICE_TYPE.getCode());
         }
+        //判断该医院该设备类型是不是被绑定过
+        HospitalequimentDTO hospitalequiment =hospitalequimentRepository.selectHospitalEquiment(hospitalCode,equipmenttypeid);
+        if (!ObjectUtils.isEmpty(hospitalequiment)){
+            throw  new IedsException(HospitalequimentEnumErrorCode.THE_DEVICE_HAS_AN_ALARM_PERIOD.getCode());
+        }
+
         hospitalequimentRepository.remove(Wrappers.lambdaQuery(new HospitalequimentPo())
         .eq(HospitalequimentPo::getHospitalcode,hospitalCode)
         .eq(HospitalequimentPo::getEquipmenttypeid,equipmenttypeid));
